+++ port/esp_idf/components/golioth_sdk/CMakeLists.txt
@@ -44,6 +44,7 @@ idf_component_register(
         "${sdk_src}/golioth_ringbuf.c"
         "${sdk_src}/golioth_event_group.c"
         "${sdk_src}/golioth_mbox.c"
+        "${sdk_src}/golioth_fw_block_processor.c"
 )
 
 component_compile_options(${EXTRA_C_FLAGS_LIST})
